refactor: Change the include path for osquery/core/plugins
