+++ osquery/core/plugins/logger.h
@@ -11,7 +11,7 @@
 
 #include <string>
 
-#include <osquery/plugins/plugin.h>
+#include <osquery/core/plugins/plugin.h>
 #include <osquery/utils/status/status.h>
 
 namespace osquery {
