+++ osquery/core/plugins/sql.h
@@ -27,9 +27,9 @@
 
 #pragma once
 
+#include <osquery/core/plugins/plugin.h>
 #include <osquery/core/sql/column.h>
 #include <osquery/core/sql/query_data.h>
-#include <osquery/plugins/plugin.h>
 #include <osquery/utils/status/status.h>
 
 namespace osquery {
