+++ osquery/experimental/events_stream/events_stream_registry.h
@@ -9,8 +9,8 @@
 
 #pragma once
 
+#include <osquery/core/plugins/plugin.h>
 #include <osquery/core/query.h>
-#include <osquery/plugins/plugin.h>
 #include <osquery/utils/expected/expected.h>
 
 #include <osquery/numeric_monitoring/numeric_monitoring.h>
