+++ osquery/logger/tests/logger.cpp
@@ -11,11 +11,11 @@
 
 #include <gtest/gtest.h>
 
+#include <osquery/core/plugins/logger.h>
 #include <osquery/core/system.h>
 #include <osquery/database/database.h>
 #include <osquery/filesystem/filesystem.h>
 #include <osquery/logger/data_logger.h>
-#include <osquery/plugins/logger.h>
 #include <osquery/registry/registry_factory.h>
 #include <osquery/utils/info/platform_type.h>
 #include <osquery/utils/system/time.h>
