+++ osquery/numeric_monitoring/plugin_interface.cpp
@@ -7,7 +7,7 @@
  * SPDX-License-Identifier: (Apache-2.0 OR GPL-2.0-only)
  */
 
-#include <osquery/plugins/plugin.h>
+#include <osquery/core/plugins/plugin.h>
 #include <osquery/registry/registry_factory.h>
 
 #include "osquery/numeric_monitoring/plugin_interface.h"
