+++ plugins/logger/aws_firehose.h
@@ -21,9 +21,8 @@
 #include <aws/firehose/model/Record.h>
 
 #include <osquery/core/core.h>
+#include <osquery/core/plugins/logger.h>
 #include <osquery/dispatcher/dispatcher.h>
-#include <osquery/plugins/logger.h>
-
 
 namespace osquery {
