+++ plugins/logger/kafka_producer.h
@@ -15,8 +15,8 @@
 #include <librdkafka/rdkafka.h>
 
 #include <osquery/core/core.h>
+#include <osquery/core/plugins/logger.h>
 #include <osquery/dispatcher/dispatcher.h>
-#include <osquery/plugins/logger.h>
 
 namespace osquery {
