+++ plugins/logger/syslog_logger.h
@@ -14,7 +14,7 @@
 #include <string>
 #include <vector>
 
-#include <osquery/plugins/logger.h>
+#include <osquery/core/plugins/logger.h>
 #include <osquery/registry/registry_factory.h>
 
 namespace osquery {
