+++ plugins/logger/tls_logger.h
@@ -11,8 +11,8 @@
 
 #include "plugins/logger/buffered.h"
 
+#include <osquery/core/plugins/logger.h>
 #include <osquery/dispatcher/dispatcher.h>
-#include <osquery/plugins/logger.h>
 
 namespace osquery {
